Added RPS - still needs tweaking. Slash commands are in seperate folder. index.js is correct file to run for bot.
